+++ cpskin/core/utils.py
@@ -157,9 +157,15 @@ def add_leadimage_from_file(container, file_name,
 def image_scale(obj, css_class, default_scale, generate_tag=True):
     images = obj.restrictedTraverse('@@images')
     if obj.portal_type in ['organization', 'person']:
-        image = images.scale('logo', scale=default_scale)
+        if getattr(images, 'logo', False):
+            image = images.scale('logo', scale=default_scale)
+        else:
+            image = None
     else:
-        image = images.scale('image', scale=default_scale)
+        if getattr(images, 'image', False):
+            image = images.scale('image', scale=default_scale)
+        else:
+            image = None
     if not image:
         return False
     if not generate_tag: